# www: improve performance of vote status endpoints