This commit focuses on improving the performance of the vote status API
endpoints.  It does so by:

1. Indexing the cast votes by token in the cache database.
2. Caching the vote results in politeiawww memory for proposals whose
   voting period has ended.

Indexing the cast votes by token in the cache database improved cast
vote lookup times by ~25%.

Most of the performance benefits come from caching the vote results in
memory for proposals whose voting period has ended.

Response times for the `/proposals/votestatus` route are included below.
The response times are the result of testing the route locally against
the pi production mainnet data.  Response times on my local machine are
significantly slower then the production site, so I've also included the
response time from the production site for a rough scale.

```
production pi - no cache : 10s

local - no cache         : 1m3s
local - cache            : 47s
local - cache & this PR  : 1.75s
```

Since only the proposals whose voting period has ended are cached in
memory, the more proposals that are actively being voted on, the slower
this route will be.  A very rough approximation is that each additional
proposal that is being actively voted on will add 0.5s to the production
site response time.  This will be addressed in a future PR.